applied each command per file
