+++ solosis/commands/farm/run_notebook.py
@@ -2,7 +2,7 @@ import os
 import subprocess
 import click
 from .. import helpers
-from solosis.utils import echo_lsf_submission_message, echo_message, log_command, bash_submit
+from solosis.utils import echo_lsf_submission_message, echo_message, log_command, bash_submit, _single_command_bsub
 
 # Script directory in the solosis package
 script_dir = os.path.dirname(os.path.abspath(__file__))  
@@ -22,8 +22,9 @@ codebase = os.path.join(script_dir, "..", "..", "codebase")
     type=str,
     help="Name of the job"
 )  
+@click.pass_context
 
-def run_ipynb(notebook, job_name, **kwargs):
+def run_ipynb(ctx, notebook, job_name, **kwargs):
     """
     Run a jupyter notebook on the farm.
     """
