+++ solosis/utils.py
@@ -107,24 +107,6 @@ def echo_lsf_submission_message(job_stdout):
 
 
 ## Functions supporting farm submission
-def bash_submit(job_runner: str, **kwargs) -> None:
-    """
-    Runs a command. Command can be a bash command (du -hs ) or a script (test.sh)
-    While running exports all kwargs as environment variables
-    This can be reused to run any bash script in all the subcommands
-    If the command needs to be submitted to farm use single_command or array_command
-    """
-    # env variables are set to
-    for k, v in kwargs.items():
-        kwargs[str(k)] = str(v)
-    # Capture result
-    result = subprocess.run(
-        [job_runner], capture_output=True, text=True, env={**os.environ, **kwargs}
-    )
-
-    click.echo(result.stdout)
-    click.echo(result.stderr)
-
 
 ## Functions supporting farm submission
 def bash_submit(job_runner: str, **kwargs) -> None:
@@ -144,3 +126,31 @@ def bash_submit(job_runner: str, **kwargs) -> None:
 
     click.echo(result.stdout)
     click.echo(result.stderr)
+
+
+def _single_command_bsub(command_to_exec, job_name, queue, time, cores, mem, **kwargs):
+    """
+    Run a single command on the farm.
+    """
+    # Script directory in the solosis package
+    script_dir = os.path.dirname(os.path.abspath(__file__))  
+    codebase = os.path.join(script_dir, "..", "..", "..", "bin")
+    job_runner = os.path.abspath(
+        os.path.join(codebase, "/farm/single_job.sh")
+        )
+    if len(command_to_exec) == 0:
+        echo_message("No command to execute", type="error")
+        return
+    
+    command_to_exec = " ".join(command_to_exec)
+    bash_submit(
+        job_runner,
+        command_to_exec=command_to_exec,
+        job_name=job_name,
+        queue=queue,
+        time=time,
+        cores=cores,
+        mem=mem,
+    )
+
+
